+++ app/routes/headers.py
@@ -13,13 +13,17 @@ header_schema = HeaderSchema()
 @jwt_required()
 def getHeaders(id=0):
     try:
+        user = get_current_user()
         project_id = get_project_id(request.args.get("project"))
-        if id==0:
-            data = HeaderModel.get_all_headers(project_id)
-            return jsonify({"headers": data}), 200, {"content-type": "application/json; charset=UTF-8"}
+        if has_access_to_project(project_id,user.id):
+            if id==0:
+                data = HeaderModel.get_all_headers(project_id)
+                return jsonify({"headers": data}), 200, {"content-type": "application/json; charset=UTF-8"}
 
-        data = HeaderModel.get_one_header(id)
-        return jsonify(data), 200, {"content-type": "application/json; charset=UTF-8"}
+            data = HeaderModel.get_one_header(id)
+            return jsonify(data), 200, {"content-type": "application/json; charset=UTF-8"}
+        else:
+            return jsonify({"Error" : "You do not have access to this project, kindly connect to project admin to access the project components"})
     except Exception as e :
         return jsonify(e), 400
 
@@ -33,31 +37,37 @@ def createHeaders():
     user = get_current_user()
     req_data['created_by'] = user.id
     req_data['modified_by'] = user.id
+    if has_access_to_project(req_data['project'],user.id):
+        try:
+            data = header_schema.load(req_data)
+        except ValidationError as err:
+            return jsonify(err), 400
 
-    try:
-        data = header_schema.load(req_data)
-    except ValidationError as err:
-        return jsonify(err), 400
+        is_exist = HeaderModel.is_exist(data.get('name'), data.get('project'))
+        
+        if is_exist:
+            return jsonify({"error": "You already have a header of the same name in this project."}), 400
 
-    is_exist = HeaderModel.is_exist(data.get('name'), data.get('project'))
-    
-    if is_exist:
-        return jsonify({"error": "You already have a header of the same name in this project."}), 400
-
-    endpoint = HeaderModel(data)
-    endpoint.save()
-    return jsonify({"success": "Header created successfully!"}), 201
+        endpoint = HeaderModel(data)
+        endpoint.save()
+        return jsonify({"success": "Header created successfully!"}), 201
+    else:
+        return jsonify({"Error" : "You do not have access to this project, kindly connect to project admin to make updates in the project components"})
 
 @headers_blueprint.route('/delete',methods=["POST"])
 @jwt_required()
 def delete_header():
     req_data = request.json
+    user = get_current_user()
     try:
         header = HeaderModel.query.get(req_data.get('header'))
     except Exception as e:
         return jsonify(str(e))
     if header:
-        header.delete()
+        if has_access_to_project(header.project,user.id):
+            header.delete()
+        else:
+            return jsonify({"Error" : "You do not have access to this project, kindly connect to project admin to make deletions in the project components"})
     else:
         return jsonify({"error" : "No such header exists"})
     return jsonify({"Success" : "Header deleted successfully"})
