+++ app/routes/payloads.py
@@ -14,13 +14,17 @@ payload_schema = PayloadSchema()
 @jwt_required()
 def get_payloads(id=0):
     try:
+        user = get_current_user()
         project_id = get_project_id(request.args.get("project"))
-        if id!=0:
-            data = PayloadModel.get_one_payload(id)
-            return jsonify(data), 200, {"content-type": "application/json; charset=UTF-8"}
+        if has_access_to_project(project_id,user.id):
+            if id!=0:
+                data = PayloadModel.get_one_payload(id)
+                return jsonify(data), 200, {"content-type": "application/json; charset=UTF-8"}
 
-        data = PayloadModel.get_all_payloads(project_id)
-        return jsonify({"payloads": data}), 200, {"content-type": "application/json; charset=UTF-8"}
+            data = PayloadModel.get_all_payloads(project_id)
+            return jsonify({"payloads": data}), 200, {"content-type": "application/json; charset=UTF-8"}
+        else:
+            return jsonify({"Error" : "You do not have access to this project, kindly connect to project admin to access the project components"})
     except Exception as e:
         return jsonify(e), 400
 
@@ -33,31 +37,37 @@ def create_payloads():
     user = get_current_user()
     req_data['created_by'] = user.id
     req_data['modified_by'] = user.id
+    if has_access_to_project(req_data['project'],user.id):
+        try:
+            data = payload_schema.load(req_data)
+        except ValidationError as err:
+            return jsonify(err), 400
 
-    try:
-        data = payload_schema.load(req_data)
-    except ValidationError as err:
-        return jsonify(err), 400
+        is_exist = PayloadModel.is_exist(data.get('name'), data.get('project'))
 
-    is_exist = PayloadModel.is_exist(data.get('name'), data.get('project'))
+        if is_exist:
+            return jsonify({"error": "You already have a payload of the same name in this project."}), 400
 
-    if is_exist:
-        return jsonify({"error": "You already have a payload of the same name in this project."}), 400
-
-    payload = PayloadModel(data)
-    payload.save()
-    return jsonify({"success": "Payload created Successfully!!"}), 201
+        payload = PayloadModel(data)
+        payload.save()
+        return jsonify({"success": "Payload created Successfully!!"}), 201
+    else:
+        return jsonify({"Error" : "You do not have access to this project, kindly connect to project admin to make updates in the project components"})
 
 @payloads_blueprint.route('/delete',methods=["POST"])
 @jwt_required()
 def delete_payload():
     req_data = request.json
+    user = get_current_user()
     try:
         payload = PayloadModel.query.get(req_data.get('payload'))
     except Exception as e:
         return jsonify(str(e))
     if payload:
-        payload.delete()
+        if has_access_to_project(payload.project,user.id):
+            payload.delete()
+        else:
+            return jsonify({"Error" : "You do not have access to this project, kindly connect to project admin to make deletions in the project components"})
     else:
         return jsonify({"error" : "No such payload exists"})
     return jsonify({"Success" : "payload deleted successfully"})
