+++ app/routes/testcases.py
@@ -14,12 +14,16 @@ testcase_schema = TestcaseSchema()
 def get_testcases(id=0):
     try:
         project_id = get_project_id(request.args.get("project"))
-        if id!=0:
-            data = TestcaseModel.get_one_testcase(id)
-            return jsonify(data), 200, {"content-type": "application/json; charset=UTF-8"}
+        user = get_current_user()
+        if has_access_to_project(project_id,user.id):
+            if id!=0:
+                data = TestcaseModel.get_one_testcase(id)
+                return jsonify(data), 200, {"content-type": "application/json; charset=UTF-8"}
 
-        data = TestcaseModel.get_all_testcases(project_id)
-        return jsonify({"testcases": data}), 200, {"content-type": "application/json; charset=UTF-8"}
+            data = TestcaseModel.get_all_testcases(project_id)
+            return jsonify({"testcases": data}), 200, {"content-type": "application/json; charset=UTF-8"}
+        else:
+            return jsonify({"Error" : "You do not have access to this project, kindly connect to project admin to access the project components"})
     except Exception as e:
         return jsonify(str(e)), 400
 
@@ -39,20 +43,22 @@ def create_testcase():
         req_data['created_by'] = user.id
         req_data['modified_by'] = user.id
         del req_data['project']
-        
-        try:
-            data = testcase_schema.load(req_data)
-        except ValidationError as err:
-            return jsonify(str(err)), 400
+        if has_access_to_project(req_data['project_id'],user.id):
+            try:
+                data = testcase_schema.load(req_data)
+            except ValidationError as err:
+                return jsonify(str(err)), 400
 
-        is_exist = TestcaseModel.is_exist(data.get('name'), data.get('project_id'))
+            is_exist = TestcaseModel.is_exist(data.get('name'), data.get('project_id'))
 
-        if is_exist:
-            return jsonify({"error": "You already have a testcases of the same name in this project."}), 400
+            if is_exist:
+                return jsonify({"error": "You already have a testcases of the same name in this project."}), 400
 
-        testcase = TestcaseModel(data)
-        testcase.save()
-        return jsonify({"success": "Testcase created successfully!"}), 201
+            testcase = TestcaseModel(data)
+            testcase.save()
+            return jsonify({"success": "Testcase created successfully!"}), 201
+        else:
+            return jsonify({"Error" : "You do not have access to this project, kindly connect to project admin to make updates in the project components"})
     except Exception as e:
         return jsonify(str(e)), 400
 
@@ -60,12 +66,16 @@ def create_testcase():
 @jwt_required()
 def delete_testcase():
     req_data = request.json
+    user = get_current_user()
     try:
         testcase = TestcaseModel.query.get(req_data.get('testcase'))
     except Exception as e:
         return jsonify(str(e))
     if testcase:
-        testcase.delete()
+        if has_access_to_project(testcase.project_id,user.id):
+            testcase.delete()
+        else:
+            return jsonify({"Error" : "You do not have access to this project, kindly connect to project admin to make deletions in the project components"})
     else:
         return jsonify({"error" : "No such testcase exists"})
     return jsonify({"Success" : "testcase deleted successfully"})
